refactor: update folder/file structure
