+++ front-end/src/components/add-user-form/AddUserForm.types.ts
@@ -1,4 +1,4 @@
-import type { User } from "../api.types";
+import { User } from "../../api/users/types";
 
 export interface AddUserFormProps {
   onSubmit: (ev: User) => void;
